Moved MixTrafficController to separate module
