+++ nym-client/src/lib.rs
@@ -1,7 +1,7 @@
 #![recursion_limit = "256"]
 
 pub mod built_info;
-pub mod clients;
+pub mod client;
 pub mod commands;
 pub mod persistence;
 pub mod sockets;
