+++ nym-client/src/persistence/pathfinder.rs
@@ -9,7 +9,7 @@ pub struct Pathfinder {
 impl Pathfinder {
     pub fn new(id: String) -> Pathfinder {
         let os_config_dir = dirs::config_dir().unwrap(); // grabs the OS default config dir
-        let config_dir = os_config_dir.join("nym").join("clients").join(id);
+        let config_dir = os_config_dir.join("nym").join("client").join(id);
         let private_mix_key = config_dir.join("private.pem");
         let public_mix_key = config_dir.join("public.pem");
         Pathfinder {
