+++ nym-client/src/sockets/ws.rs
@@ -1,5 +1,5 @@
-use crate::clients::BufferResponse;
-use crate::clients::InputMessage;
+use crate::client::BufferResponse;
+use crate::client::InputMessage;
 use directory_client::presence::Topology;
 use futures::channel::mpsc::{UnboundedReceiver, UnboundedSender};
 use futures::channel::{mpsc, oneshot};
